Time: 0 ms (100.00%), Space: 41.7 MB (74.11%) - LeetHub

+++ 1822-sign-of-the-product-of-an-array/1822-sign-of-the-product-of-an-array.java
@@ -1,7 +1,7 @@
 class Solution {
     public int arraySign(int[] nums) {
         byte temp = 1;
-        for (int i = 0; i < nums.length; i++) {
+        for (short i = 0; i < nums.length; i++) {
             if (nums[i] < 0)
                 temp *= -1;
             else if (nums[i] == 0)
